requirements.txt
tensorflow>=2.15.0
numpy
Pillow
fastapi
uvicorn[standard]
pydantic
python-multipart
psycopg2-binary